simple ui issue fixed
